Move fieldView helper to fieldView file
