+++ src/plugin/plugin.ts
@@ -8,7 +8,7 @@ import type {
   FieldNameToField,
 } from "../plugin/types/Element";
 import type { FieldNameToValueMap } from "./helpers/fieldView";
-import { getElementFieldViewFromType } from "./helpers/plugin";
+import { getElementFieldViewFromType } from "./helpers/fieldView";
 import type { Commands } from "./helpers/prosemirror";
 import { createUpdateDecorations } from "./helpers/prosemirror";
 import { getFieldNameFromNode } from "./nodeSpec";
